Move Drag and Drop to PeopleMover folder
